[CHORE/FEAT]: Seperated boot protocols into seperate modules in the kernel and basic Framebuffer driver Created a new module for the Framebuffer driver, for writing ascii text to the screen. Also cleaned up the menuconfig, as well as kernel code for the Limine boot protocol entry point.

+++ crates/limine/src/framebuffer.rs
@@ -57,7 +57,7 @@ impl<'a> Framebuffer<'a> {
         unsafe { self.inner.r0.pitch }
     }
 
-    pub fn bpp(&self) -> u64 {
+    pub fn bpp(&self) -> u16 {
         unsafe { self.inner.r0.bpp }
     }
 }
@@ -89,7 +89,7 @@ pub(crate) struct FramebufferR0 {
     height: u64,
     pitch: u64,
     /// The number of bits per pixel.
-    bpp: u64,
+    bpp: u16,
     memory_model: u8,
     red_mask_size: u8,
     red_mask_shift: u8,
